requirements.txt
tensorflow>=2.1.0
pytorch>=1.5.0
